pubmed/pyproject.toml
[project]
name = "pubmed"
version = "0.1.0"
description = ""
authors = [
    {name = "KapilNITW",email = "kskapil287@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai (>=1.5.0,<2.0.0)",
    "biopython (>=1.85,<2.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "ratelimit (>=2.2.1,<3.0.0)",
    "diskcache (>=5.6.3,<6.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "pubmedfetcher (>=0.1.0,<0.2.0)"
]

[tool.poetry]
packages = [{include = "pubmed", from = "src"}]


[tool.poetry.scripts]
get-papers-list = "pubmed.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
